feat(studioPage): 네비게이션 메뉴 확장, 축소 구현
- zustand 전역 상태 변수를 이용한 네비게이션 상태 toggle 구현
- HeaderButton 모듈화
- 네비게이션 상태를 toggle 하는 ExpandButton 모듈화

Resolves: #1
