+++ src/app/(route)/studio/[user]/_components/HeaderButton.client.tsx
@@ -1,0 +1,44 @@
+'use client';
+
+import '@/app/_styles/studioPage.css';
+import Image from 'next/image';
+
+interface ButtonProps {
+  imageSrc: string;
+  desc: string;
+  width: number;
+  height: number;
+  padding?: number;
+  onClick?: () => void;
+}
+
+const HeaderButton = ({
+  imageSrc,
+  desc,
+  width,
+  height,
+  padding,
+  onClick,
+}: ButtonProps) => {
+  return (
+    <button
+      className="transition-button relative group hover:bg-neutral-700"
+      style={{
+        padding: `${padding}px`,
+      }}
+      onClick={onClick}
+    >
+      <Image src={imageSrc} width={width} height={height} alt={desc} />
+      <span
+        className="absolute-center group-button-desc transition-button hover-opacity"
+        style={{
+          marginTop: `${(padding || 0) + 2}px`,
+        }}
+      >
+        {desc}
+      </span>
+    </button>
+  );
+};
+
+export default HeaderButton;
