+++ src/app/(route)/studio/[user]/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from './_components/Header';
+import Header from './_components/Header.server';
 
 const StudioPage = () => {
   return (
