+++ src/app/_store/studio/useNavSizeToggle.client.ts
@@ -1,0 +1,13 @@
+import { create } from 'zustand';
+
+interface Store {
+  isFold: boolean;
+  toggle: () => void;
+}
+
+const useNavSizeToggle = create<Store>()((set) => ({
+  isFold: false,
+  toggle: () => set((state) => ({ isFold: !state.isFold })),
+}));
+
+export default useNavSizeToggle;
